feat: unified retries on auth
