+++ .github/workflows/prod_fe_tests.yml
@@ -10,6 +10,14 @@ on:
 concurrency:
   group: ${{ github.workflow }}
 
+env:
+  GITHUB_TOTP: ${{ secrets.VERIFICATION_PROD_GITHUB_TOTP }}
+  GITHUB_USER: ${{ secrets.VERIFICATION_PROD_GITHUB_USER }}
+  GITHUB_PASS: ${{ secrets.VERIFICATION_PROD_GITHUB_PASS }}
+  SUPA_PLATFORM_URI: 'https://api.supabase.io/platform'
+  SUPA_META_URI: 'https://api.supabase.io/platform/pg-meta'
+  SUPA_DASHBOARD: 'https://supabase.com/dashboard'
+
 jobs:
   verification:
     name: Verification Tests
@@ -39,17 +47,11 @@ jobs:
       - name: Run verification tests
         run: npm run test:prod
         env:
-          GITHUB_TOTP: ${{ secrets.VERIFICATION_PROD_GITHUB_TOTP }}
-          GITHUB_USER: ${{ secrets.VERIFICATION_PROD_GITHUB_USER }}
-          GITHUB_PASS: ${{ secrets.VERIFICATION_PROD_GITHUB_PASS }}
           PROJECT_REF: 'nhspkocxnkjvbphkoepx'
           PROJECT_NAME: 'verifytests'
           SUPA_ORG_SLUG: 'squealing-ivory-qshl8px'
           PROJECT_DB_PASS: ${{ secrets.VERIFICATION_PROD_PROJECT_DB_PASS }}
-          SUPA_PLATFORM_URI: 'https://api.supabase.io/platform'
-          SUPA_META_URI: 'https://api.supabase.io/platform/pg-meta'
           SUPA_PROJECT_DOMAIN: 'supabase.co'
-          SUPA_DASHBOARD: 'https://app.supabase.io'
           HUBSPOT_API_KEY: ${{ secrets.VERIFICATION_PROD_HUBSPOT_API_KEY }}
 
       - uses: actions/upload-artifact@v3
@@ -87,17 +89,11 @@ jobs:
       - name: Run frontend tests
         run: npm run test:frontend:ci:prod
         env:
-          GITHUB_TOTP: ${{ secrets.VERIFICATION_PROD_GITHUB_TOTP }}
-          GITHUB_USER: ${{ secrets.VERIFICATION_PROD_GITHUB_USER }}
-          GITHUB_PASS: ${{ secrets.VERIFICATION_PROD_GITHUB_PASS }}
           FE_TESTS_PROJECT_REF: 'bwhmrdxxwjfmrelopnji'
           FE_TESTS_PROJECT_NAME: 'frontend-tests'
           FE_TESTS_PROJECT_DB_PASS: ${{ secrets.VERIFICATION_PROD_FE_TESTS_PROJECT_DB_PASS }}
           FE_TESTS_ACTOR_INFO_VIEW_ID: '29313'
-          SUPA_PLATFORM_URI: 'https://api.supabase.io/platform'
-          SUPA_META_URI: 'https://api.supabase.io/platform/pg-meta'
           SUPA_PROJECT_DOMAIN: 'supabase.co'
-          SUPA_DASHBOARD: 'https://app.supabase.com'
 
       - uses: actions/upload-artifact@v3
         if: ${{ always() }}
