+++ restore/.jest/token.ts
@@ -1,10 +1,10 @@
 import dotenv from 'dotenv'
-import { getAccessToken } from '../../src/auth/getUserToken'
+import { authenticate } from '../../src/auth/getUserToken'
 
 export default async () => {
   dotenv.config({ path: `.env.${process.env.NODE_ENV}` })
 
-  const { apiKey, contextDir } = await getAccessToken()
+  const { apiKey, contextDir } = await authenticate()
   process.env.ACCESS_TOKEN = apiKey
   process.env.CONTEXT_DIR = contextDir
 }
